Add token

+++ routes/api.php
@@ -31,13 +31,13 @@ use Illuminate\Support\Facades\Route;
 Route::controller(AuthController::class)->group(function() {
     Route::post('register', 'register');
     Route::post('login', 'login');
-    Route::post('logout', 'logout')->middleware('auth:sanctum');
+    Route::post('logout/{id}', 'logout')->middleware('auth:sanctum');
 });
 
 Route::controller(UserController::class)->group(function() {
-    Route::get('listUsers/', 'listUsers'); // Для админ панели(поиск юзера по фильтрам)
-    Route::put('updateUsers/{id}/', 'updateUsers'); // Для админ панели(изменить инфу о юзере)
-    Route::delete('deleteUsers/{id}', 'deleteUsers'); //  Для админ панели(удалить юзера)
+    Route::get('listUsers/', 'listUsers')->middleware('auth:sanctum'); // Для админ панели(поиск юзера по фильтрам)
+    Route::put('updateUsers/{id}/', 'updateUsers')->middleware('auth:sanctum');; // Для админ панели(изменить инфу о юзере)
+    Route::delete('deleteUsers/{id}', 'deleteUsers')->middleware('auth:sanctum');; //  Для админ панели(удалить юзера)
     Route::get('users/{id}', 'getUser');
     Route::get('users/{id}/social-account', 'getSocialAccountByUserId')->middleware('auth:sanctum');
     Route::post('users/favorite-event-toggle', 'toggleFavoriteEvent')->middleware('auth:sanctum');// добавляем убираем в избранное
@@ -50,7 +50,7 @@ Route::controller(AuthSocialController::class)->group(function() {
 });
 
 Route::controller(EventController::class)->group(function() {
-    Route::get('events', 'getEvents'); // Запрос ивентов с фильтрами
+    Route::get('events', 'getEvents')->middleware('auth:sanctum');; // Запрос ивентов с фильтрами
     Route::post('events/update-vk-likes', 'updateVkLikes');//для страницы мероприятия
     Route::post('events/set-event-user-liked', 'setEvenUserLiked')->middleware('auth:sanctum');//для страницы мероприятия
     Route::get('events/{id}', 'show');
@@ -79,9 +79,9 @@ Route::controller(CommentController::class)->group(function() {
 Route::controller(EventTypeController::class)->group(function() {
     Route::get('event-types', 'getTypes');
     Route::get('getTypesId/{id}', 'getTypesId');
-    Route::post('/addTypeEvent/{event_id}/{type_id}', 'addTypeEvent');
-    Route::put('/updateTypeEvent/{event_id}/{type_id}', 'updateTypeEvent');
-    Route::delete('/deleteTypeEvent/{event_id}/{type_id}', 'deleteTypeUser');
+    Route::post('/addTypeEvent/{event_id}/{type_id}', 'addTypeEvent')->middleware('auth:sanctum');;
+    Route::put('/updateTypeEvent/{event_id}/{type_id}', 'updateTypeEvent')->middleware('auth:sanctum');;
+    Route::delete('/deleteTypeEvent/{event_id}/{type_id}', 'deleteTypeUser')->middleware('auth:sanctum');;
 });
 
 Route::controller(SightTypeController::class)->group(function() {
@@ -89,24 +89,24 @@ Route::controller(SightTypeController::class)->group(function() {
 });
 
 Route::controller(StatusController::class)->group(function() {
-    Route::get('statuses', 'getStatuses');
-    Route::get('getStatusId/{id}', 'getStatusId');
-    Route::post('/addStatusEvent/{event_id}/{status_id}', 'addStatusEvent');
-    Route::put('/updateStatusEvent/{event_id}/{status_id}', 'updateStatusEvent');
-    Route::delete('/deleteStatusEvent/{event_id}/{status_id}', 'deleteStatusEvent');
+    Route::get('statuses', 'getStatuses')->middleware('auth:sanctum');
+    Route::get('getStatusId/{id}', 'getStatusId')->middleware('auth:sanctum');;
+    Route::post('/addStatusEvent/{event_id}/{status_id}', 'addStatusEvent')->middleware('auth:sanctum');;
+    Route::put('/updateStatusEvent/{event_id}/{status_id}', 'updateStatusEvent')->middleware('auth:sanctum');;
+    Route::delete('/deleteStatusEvent/{event_id}/{status_id}', 'deleteStatusEvent')->middleware('auth:sanctum');;
 
 });
 
 Route::controller(RoleController::class)->group(function() {
-    Route::get('allRole', 'allRole');
-    Route::get('getRole/{id}', 'getRole');
-    Route::post('addRole/', 'addRole');
-    Route::put('updateRole/{id}', 'updateRole');
-    Route::delete('deleteRole/{id}', 'deleteRole');
+    Route::get('allRole', 'allRole')->middleware('auth:sanctum');;
+    Route::get('getRole/{id}', 'getRole')->middleware('auth:sanctum');;
+    Route::post('addRole/', 'addRole')->middleware('auth:sanctum');;
+    Route::put('updateRole/{id}', 'updateRole')->middleware('auth:sanctum');;
+    Route::delete('deleteRole/{id}', 'deleteRole')->middleware('auth:sanctum');;
 
-    Route::post('addRoleUser/{user_id}/{role_id}', 'addRoleUser');
-    Route::put('updateRoleUser/{user_id}/{role_id}', 'updateRoleUser');
-    Route::delete('deleteRoleUser/{user_id}/{role_id}', 'deleteRoleUser');
+    Route::post('addRoleUser/{user_id}/{role_id}', 'addRoleUser')->middleware('auth:sanctum');;
+    Route::put('updateRoleUser/{user_id}/{role_id}', 'updateRoleUser')->middleware('auth:sanctum');;
+    Route::delete('deleteRoleUser/{user_id}/{role_id}', 'deleteRoleUser')->middleware('auth:sanctum');;
 });
